Mise a jour et refactorisation de la page Prestation. Déplacement et correction des import des composants
