+++ client/src/pages/Public/PrestationPage/PrestationPage.tsx
@@ -1,8 +1,8 @@
-import style from "../stylePage.module.css";
-import NavRoot from "../../components/NavRoot/NavRoot";
-import PrestationRoot from "../../components/PrestationRoot/PrestationRoot";
-import FooterRoot from "../../components/FooterRoot/FooterRoot";
-import Helmet_Prestation_SEO from "../../seo/PrestationSEO/Helmet_Prestation.seo";
+import style from "../../stylePage.module.css";
+import NavRoot from "../../../components/NavRoot/NavRoot";
+import PrestationRoot from "../../../components/Public/PrestationRoot/PrestationRoot";
+import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import Helmet_Prestation_SEO from "../../../seo/PrestationSEO/Helmet_Prestation.seo";
 
 function TarifsPage() {
      return (
